+++ addOns/spiderAjax/src/test/java/org/zaproxy/zap/extension/spiderAjax/automation/AjaxSpiderJobUnitTest.java
@@ -52,10 +52,13 @@ import org.parosproxy.paros.model.Session;
 import org.zaproxy.addon.automation.AutomationEnvironment;
 import org.zaproxy.addon.automation.AutomationJob.Order;
 import org.zaproxy.addon.automation.AutomationProgress;
+import org.zaproxy.addon.automation.AutomationStatisticTest;
 import org.zaproxy.addon.automation.ContextWrapper;
 import org.zaproxy.zap.extension.spiderAjax.AjaxSpiderParam;
 import org.zaproxy.zap.extension.spiderAjax.AjaxSpiderTarget;
 import org.zaproxy.zap.extension.spiderAjax.ExtensionAjax;
+import org.zaproxy.zap.extension.stats.ExtensionStats;
+import org.zaproxy.zap.extension.stats.InMemoryStats;
 import org.zaproxy.zap.model.Context;
 import org.zaproxy.zap.utils.I18N;
 import org.zaproxy.zap.utils.ZapXmlConfiguration;
@@ -64,6 +67,7 @@ class AjaxSpiderJobUnitTest {
 
     private static MockedStatic<CommandLine> mockedCmdLine;
     // Disdabled due to build issues
+    private ExtensionLoader extensionLoader;
     private ExtensionAjax extAjax;
     private AjaxSpiderParam ajaxSpiderParam;
 
@@ -84,12 +88,16 @@ class AjaxSpiderJobUnitTest {
 
         Model model = mock(Model.class, withSettings().defaultAnswer(CALLS_REAL_METHODS));
         Model.setSingletonForTesting(model);
-        ExtensionLoader extensionLoader = mock(ExtensionLoader.class, withSettings().lenient());
+        extensionLoader = mock(ExtensionLoader.class, withSettings().lenient());
         // extAjax = mock(ExtensionAjax.class, withSettings().lenient());
         extAjax = new ExtensionAjax();
         given(extensionLoader.getExtension(ExtensionAjax.class)).willReturn(extAjax);
         ajaxSpiderParam = new AjaxSpiderParam();
         given(extAjax.getAjaxSpiderParam()).willReturn(ajaxSpiderParam);
+
+        ExtensionStats extStats = mock(ExtensionStats.class);
+        given(extensionLoader.getExtension(ExtensionStats.class)).willReturn(extStats);
+        Mockito.lenient().when(extStats.getInMemoryStats()).thenReturn(mock(InMemoryStats.class));
 
         Control.initSingletonForTesting(Model.getSingleton(), extensionLoader);
         Model.getSingleton().getOptionsParam().load(new ZapXmlConfiguration());
@@ -130,13 +138,9 @@ class AjaxSpiderJobUnitTest {
         AjaxSpiderJob job = new AjaxSpiderJob();
 
         // When
-        job.applyCustomParameter("failIfFoundUrlsLessThan", "10");
-        job.applyCustomParameter("warnIfFoundUrlsLessThan", "11");
         job.applyCustomParameter("maxDuration", "12");
 
         // Then
-        assertThat(job.getFailIfFoundUrlsLessThan(), is(equalTo(10)));
-        assertThat(job.getWarnIfFoundUrlsLessThan(), is(equalTo(11)));
         assertThat(job.getMaxDuration(), is(equalTo(12)));
     }
 
@@ -374,24 +378,20 @@ class AjaxSpiderJobUnitTest {
     }
 
     // @Test
-    void shouldWarnIfLessUrlsFoundThanExpected() throws MalformedURLException {
+    void shouldTestAddedUrlsStatistic() throws MalformedURLException {
         // Given
-        Context context = mock(Context.class);
-        ContextWrapper contextWrapper = new ContextWrapper(context);
-
         given(extAjax.isSpiderRunning()).willReturn(false);
 
         AutomationProgress progress = new AutomationProgress();
-
-        AutomationEnvironment env = mock(AutomationEnvironment.class);
-        given(env.getDefaultContextWrapper()).willReturn(contextWrapper);
 
         AjaxSpiderJob job = new AjaxSpiderJob();
         job.setInScopeOnly(false);
 
         // When
-        job.applyCustomParameter("warnIfFoundUrlsLessThan", "1");
-        job.runJob(env, null, progress);
+        job.addTest(
+                new AutomationStatisticTest(
+                        "spiderAjax.urls.added", null, ">", 1, "warn", job.getType()));
+        job.logTestsToProgress(progress);
 
         // Then
         assertThat(job.getType(), is(equalTo("spiderAjax")));
@@ -399,37 +399,9 @@ class AjaxSpiderJobUnitTest {
         assertThat(job.getParamMethodObject(), is(extAjax));
         assertThat(job.getParamMethodName(), is("getAjaxSpiderParam"));
 
-        assertThat(progress.hasWarnings(), is(equalTo(true)));
         assertThat(progress.hasErrors(), is(equalTo(false)));
-    }
-
-    // @Test
-    void shouldErrorIfLessUrlsFoundThanExpected() throws MalformedURLException {
-        // Given
-        Context context = mock(Context.class);
-        ContextWrapper contextWrapper = new ContextWrapper(context);
-
-        given(extAjax.isSpiderRunning()).willReturn(false);
-
-        AutomationProgress progress = new AutomationProgress();
-
-        AutomationEnvironment env = mock(AutomationEnvironment.class);
-        given(env.getDefaultContextWrapper()).willReturn(contextWrapper);
-
-        AjaxSpiderJob job = new AjaxSpiderJob();
-        job.setInScopeOnly(false);
-
-        // When
-        job.applyCustomParameter("failIfFoundUrlsLessThan", "1");
-        job.runJob(env, null, progress);
-
-        // Then
-        assertThat(job.getType(), is(equalTo("spiderAjax")));
-        assertThat(job.getOrder(), is(equalTo(Order.LAST_EXPLORE)));
-        assertThat(job.getParamMethodObject(), is(extAjax));
-        assertThat(job.getParamMethodName(), is("getAjaxSpiderParam"));
-
-        assertThat(progress.hasWarnings(), is(equalTo(false)));
-        assertThat(progress.hasErrors(), is(equalTo(true)));
+        assertThat(progress.hasWarnings(), is(equalTo(true)));
+        assertThat(progress.getWarnings().size(), is(1));
+        assertThat(progress.getWarnings().get(0), is("!automation.tests.stats.fail!"));
     }
 }