+++ addOns/spiderAjax/CHANGELOG.md
@@ -9,6 +9,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Maintenance changes.
 - Handle multiple context URLs in automation.
 
+### Deprecated
+- Automation parameters `failIfFoundUrlsLessThan` and `warnIfFoundUrlsLessThan` in favour of the
+`spiderAjax.urls.added` statistic test.
+
 ## [23.3.0] - 2021-03-09
 ### Added
 - Initial support for the automation framework